requirements.txt
numpy == 1.23.3
matplotlib == 3.6.1
Pillow == 10.1.0
streamlit_image_coordinates